requirements.txt
python-telegram-bot==22.7
playwright==1.58.0
requests==2.33.0
python-dotenv>=1.0.0
psutil>=5.9.0
